requirements.txt
Flask==2.3.3
python-docx==0.8.11
gunicorn==21.2.0
